fix(src): rename do configuration
