+++ src/configuration/web/health-check.ts
@@ -1,4 +1,4 @@
-import {app} from '.';
+import {app} from './server';
 
 app.get('/health', (_, reply) => {
   void reply.status(200).send({ok: true});
