+++ src/index.ts
@@ -1,7 +1,7 @@
-import '@/infra/vendors/dotenv';
-import '@/infra/server/graphql';
-import '@/infra/server/health-check';
-import {start} from '@/infra/server';
+import '@/configuration/envs';
+import '@/configuration/web/graphql';
+import '@/configuration/web/health-check';
+import {start} from '@/configuration/web/server';
 
 console.log('Hello Starter!', {
   envs: {
